# storage/api-docs/art-info-api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Art-UA Info API",
        "description": "API для інформаційного порталу art-ua.info — новини, події, статті про мистецтво",
        "contact": {
            "name": "Art-UA Info Support",
            "email": "info@art-ua.info"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://art-ua.info/api/v1",
            "description": "Production Server"
        },
        {
            "url": "https://save-art.ddev.site/api/v1",
            "description": "Local Development"
        }
    ],
    "paths": {
...
        "/v1/projects/{project}/like": {
            "post": {
                "tags": [
                    "Likes"
                ],
                "summary": "Поставити лайк",
                "description": "Додає лайк до проекту від поточного користувача",
                "operationId": "likeProject",
                "parameters": [
                    {
                        "name": "project",
                        "in": "path",
                        "description": "ID проекту",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Лайк додано",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "is_liked": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "likes_count": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Вже лайкнуто"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Likes"
                ],
                "summary": "Прибрати лайк",
                "description": "Видаляє лайк з проекту",
                "operationId": "unlikeProject",
                "parameters": [
                    {
                        "name": "project",
                        "in": "path",
                        "description": "ID проекту",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Лайк прибрано",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "is_liked": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "likes_count": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Не було лайка"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/v1/catalogs/{artCatalog}/like": {
            "post": {
                "tags": [
                    "Likes"
                ],
                "summary": "Поставити лайк каталогу",
                "description": "Додає лайк до PDF-каталогу від поточного користувача",
                "operationId": "likeArtCatalog",
                "parameters": [
                    {
                        "name": "artCatalog",
                        "in": "path",
                        "description": "ID каталогу",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Лайк додано",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "is_liked": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "likes_count": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Вже лайкнуто"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Likes"
                ],
                "summary": "Прибрати лайк каталогу",
                "description": "Видаляє лайк з PDF-каталогу",
                "operationId": "unlikeArtCatalog",
                "parameters": [
                    {
                        "name": "artCatalog",
                        "in": "path",
                        "description": "ID каталогу",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Лайк прибрано",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "is_liked": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "likes_count": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Не було лайка"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum Bearer Token. Формат: Bearer <token>",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            },
            "apiKey": {
                "type": "apiKey",
                "description": "API ключ для доступу до API. Передається у заголовку X-Api-Key.",
                "name": "X-Api-Key",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "apiKey": []
        }
    ]
}